requirements.txt
streamlit==1.32.0
psycopg2-binary==2.9.9
groq==0.9.0
httpx==0.27.2
pymupdf==1.24.0
python-docx==1.1.0
plotly==5.20.0
pandas==2.2.1
bcrypt==4.1.2
python-dotenv==1.0.1
Pillow==10.2.0
